requirements.txt
discord.py==2.4.0
aiohttp==3.10.5
beautifulsoup4==4.12.3
lxml==5.3.0
yarl==1.9.4
